+++ nix/package.nix
@@ -14,7 +14,7 @@
 
 buildPythonApplication {
   pname = "iaclens";
-  version = "0.6.0";  # keep in sync with pyproject.toml
+  version = "0.7.0";  # keep in sync with pyproject.toml
   pyproject = true;
   src = lib.cleanSource ../.;